VideoToScreenshots/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
fonttools==4.48.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
OpenCC==0.2
opencv-python==4.9.0.80
packaging==23.2
pillow<=9.5.0
PyAutoGUI==0.9.54
pydantic==2.6.1
pydantic_core==2.16.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pysrt==1.1.2
python-dateutil==2.8.2
pytweening==1.0.7
regex==2023.12.25
requests==2.31.0
rubicon-objc==0.4.7
six==1.16.0
sniffio==1.3.0
srt==3.5.3
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
